==> 03_mr_mediators_to_BC.Rmd <==
---
title: "Univariate MR of mediators to BC"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(readr)
library(data.table)
library(vroom)

library(tidyr)
library(purrr)
library(tibble)
library(dplyr)

library(TwoSampleMR)
library(ggplot2)
library(cowplot)
library(wesanderson)
```

```{r message=F}
# set path for pre-calculated data, outside the code repo
# `local` / `remote` (reading data from RDSF)
currently_working_env = "local"
source("set_paths.R")
set_paths(currently_working_env)

# metafile
data_lookup<-read_csv(paste0("metadata/data_lookup.csv")) 
bcac_lookup<-read_csv(paste0("metadata/data_lookup_BCAC.csv")) %>% filter(!grepl("BRCA1", trait))

# load functions
source("functions.R")
```


```{r}
# specify group to process if data is in textfiles
source = "textfiles"
current_trait_category <- "md_unadj"
mediators <- data_lookup %>% filter(trait_category == current_trait_category) %>% filter(!is.na(full_data)) %>%  pull(tophits_data)

# specify group to process  if the data is in MRBase
#source = "mrbase"
#current_trait_category <- ".."
#mediators <- data_lookup %>% filter(trait_category == current_trait_category) %>% filter(!is.na(mrbase.id)) %>%  pull(mrbase.id)

```


## Reading in / extracting instruments
```{r}
if (source == "textfiles"){
  
    data_list <- lapply(paste0(data_path_tophits, mediators), fread)
    
} else if (source == "mrbase"){

  data_list <- list()
  for (i in 1:length(mediators)){
    instruments <- extract_instruments(mediators[i])
    if (is.null(instruments)){ stop(paste0("No instruments found for ", mediators[i]))} # if no SNPs returned
    data_list[[i]] <- clump_data(instruments)
    data_list[[i]]$exposure <- data_lookup %>% filter(mrbase.id==mediators[i]) %>% pull(trait)
  }
}
```




```{r message=F, warning=F}

extract_outcome_data_custom <- function(exposure_dat, breast_cancer_data, outcome_source){
  # function to extract instrument from a specified outcome
  
  if (outcome_source == "outcome_textfile"){
    out <- breast_cancer_data %>%
              filter(SNP %in% exposure_dat$SNP)
    
  } else if (outcome_source == "outcome_mrbase"){
  
    out <- extract_outcome_data(
              snps = exposure_dat$SNP,
              outcome = breast_cancer_data) 
  }
  return(out)
}
```


```{r}
# This for loop will iterate over all breast cancer outcomes and do all steps of MR

for (i in 1:length(bcac_lookup$trait)){

  if (!is.na(bcac_lookup$mr.base.id[i])){
    outcome_source = 'outcome_mrbase'
    outcome_name = bcac_lookup$mr.base.id[i]
    print(paste0( "Testing all mediators against breast cancer data: ", outcome_name))
    breast_cancer_data <- outcome_name
  } else {
    outcome_source = 'outcome_textfile'
    outcome_name = bcac_lookup$full_data[i]
    print(paste0( "Testing all mediators against breast cancer data: ", outcome_name))
    # read in the outcome file
    breast_cancer_data  <- vroom(paste0(data_path_gwas, outcome_name)) 
  }
  outcome_trait_name = bcac_lookup$trait_file_name[i]
  outcome_sample_size =  bcac_lookup$sample.size[i]
  outcome_ncase = bcac_lookup$ncase[i]
  outcome_ncontrol = bcac_lookup$ncontrol[i]

  outcomes <- lapply(data_list, extract_outcome_data_custom, breast_cancer_data, outcome_source)
  
  # check how many SNPs are present in the outcome
  for (i in 1:length(data_list)){
    print(unique(data_list[[i]]$exposure))
    print(paste0("SNPs in exposure: ", dim(data_list[[i]])[1]))
    print(paste0("SNPs matches outcome: ", dim(outcomes[[i]])[1]))
  }
  
  # Harmonise each exposure to the matched outcome SNPs
  harmonise_l <- list()
  for (i in 1:length(data_list)){
     harmonise_l[[i]] <- harmonise_data(exposure_dat = data_list[[i]], 
                                        outcome_dat = outcomes[[i]])
  }
  
  steiger_l <- list()
  for (i in 1:length(data_list)){
     steiger_l[[i]] <- calc_steiger(harmonise_l[[i]], 
                                    exposure_ss = 24192,
                                    outcome_ss = outcome_sample_size, 
                                    outcome_ncase = outcome_ncase,
                                    outcome_ncontrol = outcome_ncase)$summary
  }
  
  # Pleiotropy and Heterogeneity
  sensitivity_l <- list()
  for (i in 1:length(harmonise_l)){
    # don't try it one 1-SNP cases
    if (dim(harmonise_l[[i]])[1] != 1){
    sensitivity_l[[i]] <- full_join(
                          mr_pleiotropy_test(harmonise_l[[i]]),
                          mr_heterogeneity(harmonise_l[[i]], method_list=c("mr_egger_regression", "mr_ivw"))
                          )
    }
  }   
    
  # perform MR
  mr_l <- list()
  for (i in 1:length(harmonise_l)){
    mr_l[[i]] <- mr(harmonise_l[[i]],  method_list=c('mr_ivw','mr_egger_regression','mr_weighted_median', 'mr_wald_ratio') ) 
  }
  
  
  ## Tidy up results 
  #  - Split outcome names
  #  - Generate odds ratios with 95% confidence intervals
    
  tidy_result_l <- list()
  for (i in 1:length(mr_l)){
    tidy_result_l[[i]] <- mr_l[[i]] %>%
                          split_outcome() %>% 
                          split_exposure() %>% 
                          separate(outcome, "outcome", sep="[(]") %>% 
                          generate_odds_ratios()
  }
  
  dat <- tidy_result_l %>%
        reduce(rbind) %>% 
        arrange(method) %>% 
        select(-c("id.exposure")) 
  
  # save full data from MR analyses
  ind_res_dir <- paste0(results_path, current_trait_category, "/")
  ind_res_dir_w_backslash <- gsub(" ", "\\ ", ind_res_dir, fixed=T) # create path vector escaping spaces, otherwise sytem call cant process it
  if(!dir.exists(ind_res_dir)){ system(paste("mkdir -p", ind_res_dir_w_backslash))}
  
  
  write_tsv(dat, paste0(ind_res_dir,"merged_mr_mediators_", current_trait_category,"-to-BCAC_", outcome_trait_name,".tsv"))
  
  
  dat_sensitivity <- sensitivity_l %>%
        reduce(rbind) %>% 
        split_outcome() %>%
        separate(outcome, "outcome", sep="[(]") %>% 
        select(-c("id.exposure")) 
  
  dat_steiger <- steiger_l %>% 
        reduce(rbind) %>% 
        separate(outcome, "outcome", sep="[(]") 
  
  dat_sensitivity <- left_join(dat_sensitivity, dat_steiger, by = c("exposure" = "exposure", "outcome"="outcome"))
  
  
  
  # save sensitivity analyses results
  write_tsv(dat_sensitivity, paste0(ind_res_dir,"merged_sens_mr_mediators_", current_trait_category,"-to-BCAC_", outcome_trait_name,".tsv"))
  
}
```

# TESTING
```{r}
da_old <- read_tsv(paste0(data_path_tophits, "mammographic_density_DA_sieh_tophits.tsv")) #16
nda_old <- read_tsv(paste0(data_path_tophits, "mammographic_density_NDA_sieh_tophits.tsv")) # 11
pd_old <- read_tsv(paste0(data_path_tophits, "mammographic_density_PD_sieh_tophits.tsv")) # 10
                   

da_new <- read_tsv(paste0(data_path_tophits, "dense_area_tophits.tsv")) # 25
nda_new <- read_tsv(paste0(data_path_tophits, "nondense_area_tophits.tsv")) # 16
pd_new <- read_tsv(paste0(data_path_tophits, "percent_density_tophits.tsv")) # 16

da_unadj <- read_tsv(paste0(data_path_tophits, "dense_area_unadj_tophits.tsv")) # 21
nda_unadj <- read_tsv(paste0(data_path_tophits, "nondense_area_unadj_tophits.tsv")) # 8
pd_unadj <- read_tsv(paste0(data_path_tophits, "percent_density_unadj_tophits.tsv")) # 11


length(intersect(da_old$SNP , da_new$SNP)) # 15
length(intersect(nda_old$SNP , nda_new$SNP)) #11
length(intersect(pd_old$SNP , pd_new$SNP)) # 9

length(intersect(da_unadj$SNP , da_new$SNP)) # 8
length(intersect(nda_unadj$SNP , nda_new$SNP)) #1
length(intersect(pd_unadj$SNP , pd_new$SNP)) # 1


md_gwas_DA_tophits <- read_csv(paste0(data_path, "MD_raw_files/", "MD_significant_snps_DA_2022-02-11.csv")) %>%  # 668
                      select(-chrom, -pos, SNP = rsid)
md_gwas_NDA_tophits <- read_csv(paste0(data_path, "MD_raw_files/", "MD_significant_snps_NDA_2022-02-11.csv")) %>% # 670
                      select(-chrom, -pos, SNP = rsid)
md_gwas_PD_tophits <- read_csv(paste0(data_path, "MD_raw_files/", "MD_significant_snps_PD_2022-02-11.csv")) %>%  # 484
                      select(-chrom, -pos, SNP = rsid)

 
length(intersect(md_gwas_DA_tophits$SNP , da_new$SNP)) # 26 # OK
length(intersect(md_gwas_NDA_tophits$SNP , nda_new$SNP)) #16
length(intersect(md_gwas_PD_tophits$SNP , pd_new$SNP)) # 16

length(intersect(md_gwas_DA_tophits$SNP , da_old$SNP)) # 16  # OK
length(intersect(md_gwas_NDA_tophits$SNP , nda_old$SNP)) #11
length(intersect(md_gwas_PD_tophits$SNP , pd_old$SNP)) # 10






full_join(
  da_old %>%  dplyr::select(SNP, old_beta.exposure = beta.exposure, old_pval = pval.exposure),
  da_new %>%  dplyr::select(SNP, new_beta.exposure = beta.exposure, new_pval = pval.exposure),
  by ='SNP') %>% mutate(prev = ifelse(SNP %in% prev_reported, T, F))  %>% View()




md_gwas_DA_tophits <- read_csv(paste0(data_path, "MD_raw_files/", "MD_significant_snps_DA_2022-02-11.csv")) %>% 
                      dplyr::select(-chrom, -pos, SNP = rsid)


prev_reported <- c(
"rs11205277", "rs11205303", 
"rs12468790", "rs4849864", 
"rs17625845", "rs10034692", 
"rs149689338", "rs71219402", 
"rs186749", "rs335143", 
"rs335160", "rs9485370", 
"rs4897108", "rs12665607", 
"rs60705924", "rs7816345", 
"rs10086016", "rs10995190", 
"rs10995181", "rs2089176", 
"rs3817198", "rs703556", 
"rs5800526", "rs1265507", 
"rs9302903", "rs7289126", 
"rs17001868", "rs73169057", 
"rs6001984")

md_gwas_DA_tophits %>% mutate(prev = ifelse(SNP %in% prev_reported, T, F)) %>% View()


## extract from adj and unadjust

# da
da_new_all <- vroom("/Users/ny19205/OneDrive - University of Bristol/Documents - OneDrive/Mini-project4/01_Data/GWAS_tidy/dense_area_GWAS_tidy_outcome.txt.gz") %>% select(SNP, beta_adj = beta.outcome, pval_adj = pval.outcome)
da_unadj_all <- vroom("/Users/ny19205/OneDrive - University of Bristol/Documents - OneDrive/Mini-project4/01_Data/GWAS_tidy/dense_area_unadj_GWAS_tidy_outcome.txt.gz")%>% select(SNP, beta_unadj = beta.outcome, pval_unadj = pval.outcome)

da_merged<- full_join(da_new_all, da_unadj_all) %>% 
            filter(grepl("rs", SNP)) %>% 
            mutate(adj = ifelse(SNP %in% da_new$SNP, 1, 0)) %>% 
            mutate(unadj = ifelse(SNP %in% da_unadj$SNP, 1, 0)) %>% 
            filter(adj==1 | unadj==1)
write_tsv(da_merged, paste0(results_path, current_trait_category, "/compare_DA_adj_unadj.tsv"))
da_merged<- read_tsv(paste0(results_path, current_trait_category, "/compare_DA_adj_unadj.tsv"))

# nda
nda_new_all <- vroom("/Users/ny19205/OneDrive - University of Bristol/Documents - OneDrive/Mini-project4/01_Data/GWAS_tidy/nondense_area_GWAS_tidy_outcome.txt.gz") %>% select(SNP, beta_adj = beta.outcome, pval_adj = pval.outcome)
nda_unadj_all <- vroom("/Users/ny19205/OneDrive - University of Bristol/Documents - OneDrive/Mini-project4/01_Data/GWAS_tidy/nondense_area_unadj_GWAS_tidy_outcome.txt.gz")%>% select(SNP, beta_unadj = beta.outcome, pval_unadj = pval.outcome)

nda_mergedX<- full_join(nda_new_all, nda_unadj_all) %>% 
            filter(grepl("rs", SNP)) %>% 
            mutate(adj = ifelse(SNP %in% nda_new$SNP, 1, 0)) %>% 
            mutate(unadj = ifelse(SNP %in% nda_unadj$SNP, 1, 0)) %>% 
            filter(adj==1 | unadj==1)
write_tsv(nda_merged, paste0(results_path, current_trait_category, "/compare_NDA_adj_unadj.tsv"))
nda_merged<- read_tsv(paste0(results_path, current_trait_category, "/compare_NDA_adj_unadj.tsv"))


# pd
pd_new_all <- vroom("/Users/ny19205/OneDrive - University of Bristol/Documents - OneDrive/Mini-project4/01_Data/GWAS_tidy/percent_density_GWAS_tidy_outcome.txt.gz") %>% select(SNP, beta_adj = beta.outcome, pval_adj = pval.outcome)
pd_unadj_all <- vroom("/Users/ny19205/OneDrive - University of Bristol/Documents - OneDrive/Mini-project4/01_Data/GWAS_tidy/percent_density_unadj_GWAS_tidy_outcome.txt.gz")%>% select(SNP, beta_unadj = beta.outcome, pval_unadj = pval.outcome)

pd_merged<- full_join(pd_new_all, pd_unadj_all) %>% 
            filter(grepl("rs", SNP)) %>% 
            mutate(adj = ifelse(SNP %in% pd_new$SNP, 1, 0)) %>% 
            mutate(unadj = ifelse(SNP %in% pd_unadj$SNP, 1, 0)) %>% 
            filter(adj==1 | unadj==1)
write_tsv(pd_merged, paste0(results_path, current_trait_category, "/compare_PD_adj_unadj.tsv"))
pd_merged<- read_tsv(paste0(results_path, current_trait_category, "/compare_PD_adj_unadj.tsv"))

```
