==> 03_perform_MR_for_BMI_to_MD.Rmd <==
---
title: "Univarite MR of BMI to MD"
author: "Marina Vabistsevits"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(readr)
library(data.table)
library(vroom)

library(tidyr)
library(purrr)
library(tibble)
library(dplyr)

library(TwoSampleMR)
library(ggplot2)
library(cowplot)
library(wesanderson)
```

```{r}
# set path for pre-calculated data, outside the code repo
# `local` / `remote` (reading data from RDSF)
currently_working_env = "local"
source("set_paths.R")
set_paths(currently_working_env)

# metafile
data_lookup<-read_csv(paste0("metadata/data_lookup.csv")) 

# load functions
source("functions.R")
```

# Load BMI exposure files
```{r message=F}

early_bmi_file <- data_lookup %>% filter(trait == "Childhood body size") %>% pull(tophits_data)
early_bmi_exp <- read_tsv(paste0(data_path_tophits, early_bmi_file))
dim(early_bmi_exp) # 115 

adult_bmi_file <- data_lookup %>% filter(trait == "Adult body size") %>% pull(tophits_data)
adult_bmi_exp <- read_tsv(paste0(data_path_tophits, adult_bmi_file))
dim(adult_bmi_exp) # 173 
```



# Run MR for all 

```{r}
# specify group to process if data is in textfiles
current_trait_category <- "md_unadj"
mediators <- data_lookup %>% filter(trait_category == current_trait_category) %>% filter(!is.na(full_data)) %>%  pull(full_data) 
```

```{r message =F}
# MR loop
for (i in 1:length(mediators)){
  
  # Load data (first decide if it's in text file or in MR base)
  
  if( mediators[i] %in% data_lookup$full_data ){
    # it's a text file
    mediator_file_name <- data_lookup %>% filter(full_data == mediators[i]) %>% pull(trait_file_name)
    
    # load full GWAS data (outcomes) and subset to exposure SNPs
    print(paste0("Processing ", mediators[i]))
    outcome_mediator <- vroom(paste0(data_path_gwas, mediators[i]))
  
    outcome_mediator_early <- outcome_mediator %>%
              filter(SNP %in% early_bmi_exp$SNP)
    outcome_mediator_adult <- outcome_mediator %>%
              filter(SNP %in% adult_bmi_exp$SNP)
    rm(outcome_mediator)
    

  } else if ( mediators[i] %in% data_lookup$mrbase.id ){
    # it's in mrbase
    mediator_file_name <- data_lookup %>% filter(mrbase.id == mediators[i]) %>% pull(trait_file_name)

   # extrarct exposure SNPs from outcomes in (mediators) in MRBase
   print(paste0("Processing ", mediators[i]))
   outcome_mediator_early <- extract_outcome_data(
              snps = early_bmi_exp$SNP,
              outcome = mediators[i])
   outcome_mediator_adult <- extract_outcome_data(
              snps = adult_bmi_exp$SNP,
              outcome = mediators[i])
   # add better outcome names
   outcome_mediator_early$outcome <- data_lookup %>% filter(mrbase.id == mediators[i]) %>% pull(trait)
   outcome_mediator_adult$outcome <- data_lookup %>% filter(mrbase.id == mediators[i]) %>% pull(trait)
  }
  
    
  # Harmonise 
  dat_early <- harmonise_data(exposure_dat = early_bmi_exp, 
                          outcome_dat = outcome_mediator_early)
  dat_adult <- harmonise_data(exposure_dat = adult_bmi_exp, 
                          outcome_dat = outcome_mediator_adult)
  
  streiger_early =  calc_steiger(dat_early, exposure_ss = 246511, outcome_ss = 24192) 
  streiger_adult =  calc_steiger(dat_adult, exposure_ss = 206658, outcome_ss = 24192) 


  # Perform MR
  res_early <- mr(dat_early, method_list=c('mr_ivw','mr_egger_regression','mr_weighted_median', 'mr_wald_ratio')) 
  res_adult <- mr(dat_adult, method_list=c('mr_ivw','mr_egger_regression','mr_weighted_median', 'mr_wald_ratio'))   
  
  
  # exploratory
  #res_single <- mr_singlesnp(dat_early, all_method=c("mr_ivw_mre","mr_egger_regression", "mr_weighted_median"))%>% 
  #                          mutate(SNP = gsub("Inverse variance weighted (multiplicative random effects)", "IVW (MRE)", SNP, fixed = T))
  #steiger_wrong<- streiger_early$single_rsq %>% filter(steiger_dir == F) %>% pull(SNP)
  #single_forest_cols <- mr_forest_plot_outliers(res_single,  outliers_list=steiger_wrong,
  #                                              outliers_colour = "tomato", method = "Steiger", or=F) 
  #ggsave(plot=single_forest_cols, 
  #    height=15, width=5,
  #    filename=paste0(results_path,  "/testing_steiger_BMItoMD/earlyBMI_to_", mediator_file_name,"_snps_steiger_dir.png"))
#

  #res_early_steiger <- mr(dat_early %>% filter(!SNP %in% steiger_wrong),
  #                method_list=c('mr_ivw','mr_egger_regression','mr_weighted_median', 'mr_wald_ratio')) 
  #res_early_steiger_tidy<- res_early_steiger %>%
  #            split_outcome() %>% 
  #            split_exposure() %>% 
  #            generate_odds_ratios()

  #res_early_tidy$instruments <- 'tophits'
  #res_early_steiger_tidy$instruments <- "steiger filtered"
  #pd_compare <- bind_rows(res_early_tidy,res_early_steiger_tidy )
  #compare_all<- bind_rows(da_compare, nda_compare, pd_compare) %>% 
  #  select(exposure, outcome, method, nsnp, instruments, b,lo_ci, up_ci, se,pval)
  #compare_all %>%  write_csv(paste0(results_path,  "/testing_steiger_BMItoMD/steiger_mr_bmi_to_md.csv"))
          
  # Tidy up results and save
  res_early_tidy<- res_early %>%
                split_outcome() %>% 
                split_exposure() %>% 
                generate_odds_ratios()
  
  res_adult_tidy<- res_adult %>%
                split_outcome() %>% 
                split_exposure() %>%
                generate_odds_ratios()
  
  out <- rbind(res_early_tidy, res_adult_tidy ) %>% 
          arrange(method) %>%
          select(-c("id.outcome", "id.exposure"))
  
  
  # Pleiotropy and Heterogeneity
  out_sensitivity<- rbind(
    full_join(mr_pleiotropy_test(dat_early),
              mr_heterogeneity(dat_early, method_list=c("mr_egger_regression", "mr_ivw"))) %>% 
      left_join(streiger_early$summary),
    full_join(mr_pleiotropy_test(dat_adult),
              mr_heterogeneity(dat_adult, method_list=c("mr_egger_regression", "mr_ivw"))) %>% 
      left_join(streiger_adult$summary)
  )  %>% select(-c("id.outcome", "id.exposure"))       
  
  
  ind_res_dir <- paste0(results_path, current_trait_category, "/")
  ind_res_dir_w_backslash <- gsub(" ", "\\ ", ind_res_dir, fixed=T) # create path vector escaping spaces, otherwise sytem call cant process it
  if(!dir.exists(ind_res_dir)){ system(paste("mkdir -p", ind_res_dir_w_backslash))}
  
  write_tsv(out, paste0(ind_res_dir, "MR_BMI-to-", mediator_file_name ,".tsv"))
  write_tsv(out_sensitivity, paste0(ind_res_dir, "sens_MR_BMI-to-", mediator_file_name ,".tsv"))
  print(paste0("Finished MR for ", mediator_file_name)) 
}

```


### Saving individual traits into a merged table + subset to IVW only -- this will be used for plotting

```{r}
## save results for one category into one df

# load all results files into a table to view together
current_trait_category <- "md_unadj" ###CHANGE ME
ind_res_dir <- paste0(results_path, current_trait_category)

## MR results
all.files_mr <- list.files(path = ind_res_dir, pattern = "MR_BMI-to-*", full.names = T)
all.files_mr <- all.files_mr[!grepl('sens',all.files_mr)]

l <- lapply(all.files_mr, fread)
df <- l %>% reduce(rbind)
write_tsv(df, paste0(results_path, current_trait_category, "/","merged_mr_BMI-to-", current_trait_category,".tsv"))


## Sensitivity analysis results
all.files_mr <- list.files(path = ind_res_dir, pattern = "sens_MR_BMI-to-*", full.names = T)

l <- lapply(all.files_mr, fread)
df <- l %>% reduce(rbind) 

write_tsv(df, paste0(results_path, current_trait_category, "/merged_sens_mr_BMI-to-", current_trait_category,".tsv"))
```
