==> 05_mvmr_create_plots.Rmd <==
---
title: "MVMR make plots"
author: "Marina Vabistsevits"
date: "01/06/2020"
output: html_document
---

```{r setup, include=FALSE}
library(cowplot)
library(data.table)
library(dplyr)
library(ggplot2)
library(readr)
library(tidyr)
library(TwoSampleMR)
library(wesanderson)
knitr::opts_chunk$set(echo = TRUE)
```

```{r message=F}
# set path for pre-calculated data, outside the code repo
# `local` / `remote` (reading data from RDSF)
currently_working_env = "local"
source("set_paths.R")
set_paths(currently_working_env)

# metafile
data_lookup<-read_csv(paste0("metadata/data_lookup.csv")) 
source("functions.R")

```

```{r}
# specify category and palette to use

current_trait_category <- "md_unadj"
pal<-c(unname(yarrr::piratepal("pony")))
```

## MVMR analysis 1: childhood BMI + adult BMI -> mediator

```{r message=F}
dat <- read_tsv(paste0(results_path, current_trait_category, 
                       "/mvmr/merged/merged_mvmr_BMIs-", current_trait_category,"-to-BCAC.tsv"))


pal<-c(unname(yarrr::piratepal("pony")))
p2<-ggplot(dat, aes(y=exposure, x=b, label=outcome, shape=exposure, colour=exposure)) +
  geom_errorbarh(aes(xmin=lo_ci, xmax=up_ci), height=0.2) +
  geom_point(size=2)+
  #xlim(-1.2,0.5)+
  scale_color_manual(values=pal)+
  scale_shape_manual(values = c(19,20)) +
  geom_vline(xintercept=0, linetype=3,) +
  theme_minimal_hgrid(10, rel_small = 1) +
  facet_wrap(~outcome, ncol=1)+
  labs(color = "",y = "", x = "Beta",
       title= paste0("MVMR results for ", gsub("_", " ",current_trait_category) ,", 95% CI") )+
  theme(legend.position = "none", plot.title.position  = "plot")

ggsave(paste0("figures/BMI-to-",current_trait_category ,"_MVMR.png"),
       plot=p2, scale=2, 
       width=10, height=5,
       units=c("cm"), dpi=200, limitsize=F)

```

```{r}
x<-plot_grid(p, p2, labels = c('', ''), nrow=2) # p is from 02_...Rmd

ggsave(paste0("figures/BMI-to-MD_unimr_mvmr.png"),
       plot=x, scale=1.4, bg="white",
       width=8, height=11,
       units=c("cm"), dpi=300, limitsize=F)
```



## MVMR analysis 2: childhood BMI + mediator -> breast cancer

```{r}
# previous results
dat_BMI <-  read_tsv("/Users/ny19205/OneDrive - University of Bristol/Documents - OneDrive/Mini-project2/early-bmi-breast-cancer-mr/figures/plotting_data/fig4.tsv")

pp<-ggplot(dat_BMI, aes(y=reorder(direct, -or), x=or, label=outcome, colour=text_col)) +
  geom_errorbarh(aes(xmin=or_lci95, xmax=or_uci95), height=.3) +
  geom_point(size=2)+
  geom_vline(xintercept=0.6632422, linetype=2, colour = "steelblue") + # cBMI total effect
  geom_vline(xintercept=1, linetype='longdash') +
  theme_minimal_hgrid(10) +
  #scale_color_manual(values=c("tomato",  "steelblue"))+# old
  scale_color_manual(values=c("tomato",  "orange","steelblue"))+ # with MD
  scale_y_discrete(position = "right")+
  theme(strip.text = element_text(face = 'bold'))+
  facet_grid(~outcome, scale="free", switch="x")+
  labs(color = "",y = "", x = "Odds ratio", subtitle="" )+
  theme(legend.position = "none")


```


```{r message=F}

dat <- read_tsv( paste0(results_path, current_trait_category, "/mvmr/merged/merged_mvmr_earlyBMI-",current_trait_category,"-to-BCAC.tsv")) %>% 
  mutate(mediator = case_when(grepl("nondense_area", mediator) ~ "MD non-dense area (unadj)",
                              grepl("dense_area", mediator) ~ "MD dense area (unadj)",
                              grepl("percent_density", mediator) ~ "MD percent density (unadj)")) %>% 
  filter(outcome == "Breast cancer") %>% 
  filter(exposure == "Childhood body size") %>% 
  mutate(direct = paste0(exposure, " accounted for ", mediator)) %>% 
  select(direct, outcome, starts_with("or")) %>% 
  mutate(effect = 'direct', text_col="orange")

dat_BMI <- bind_rows(dat_BMI, dat)

# + rerun prev block!

```
